+++ R/as-discretised-scale.R
@@ -1,0 +1,36 @@
+#' Create discretised versions of continuous scales
+#'
+#' @param scale_funcion a scale function (e.g. `scale_fill_divergent`)
+#'
+#'
+#' @return
+#' A function with the same arguments as `sacle_function` that works with discreteised
+#' values
+#'
+#' @examples
+#' library(ggplot2)
+#' scale_fill_brewer_discretised <- as.discretised_scale(scale_fill_distiller)
+#'
+#' @seealso scale_fill_discreteised
+#'
+#' @export
+as.discretised_scale <- function(scale_function) {
+    new_fun <- function(...) {
+        call <- match.call()
+
+        if (is.null(call$super)) {
+            call$super <- ScaleDiscretised
+        }
+
+        if (is.null(call$guide)) {
+            call$guide <- ggplot2::guide_colorsteps()
+        }
+
+        call[[1]] <- scale_function
+
+        eval(call)
+    }
+
+    formals(new_fun) <- formals(scale_function)
+    new_fun
+}